[#1616] Move reactive tunnel tests into conformance test suite
